pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "driver-api-doc"
version = "1.0.0"
description = "智能化驱动API文档生成工具"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Driver API Doc Agent"}
]
keywords = ["driver", "api", "documentation", "generator", "embedded", "c"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    "pyyaml>=6.0",
    "anthropic>=0.18.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0",
]

[project.scripts]
driver-doc = "agent.main:main"

[project.urls]
Homepage = "https://github.com/example/driver-api-doc"
Documentation = "https://github.com/example/driver-api-doc#readme"
Repository = "https://github.com/example/driver-api-doc"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*"]

[tool.setuptools.package-data]
agent = ["py.typed"]
